refactor(libdd-capabilities): remove transitive dependency
