+++ datadog-sidecar/Cargo.toml
@@ -18,6 +18,7 @@ anyhow = { version = "1.0" }
 arrayref = "0.3.7"
 priority-queue = "2.1.1"
 libdd-common = { path = "../libdd-common" }
+libdd-capabilities = { path = "../libdd-capabilities", version = "0.1.0" }
 datadog-sidecar-macros = { path = "../datadog-sidecar-macros" }
 
 libdd-telemetry = { path = "../libdd-telemetry", features = ["tracing"] }
