+++ datadog-sidecar/src/service/tracing/trace_flusher.rs
@@ -5,8 +5,9 @@ use super::TraceSendData;
 use crate::agent_remote_config::AgentRemoteConfigWriter;
 use datadog_ipc::platform::NamedShmHandle;
 use futures::future::join_all;
-use libdd_common::capabilities::HttpClientTrait;
-use libdd_common::{DefaultHttpClient, Endpoint, MutexExt};
+use libdd_capabilities::HttpClientTrait;
+use libdd_capabilities_impl::DefaultHttpClient;
+use libdd_common::{Endpoint, MutexExt};
 use libdd_trace_utils::trace_utils;
 use libdd_trace_utils::trace_utils::SendData;
 use libdd_trace_utils::trace_utils::SendDataResult;
