+++ libdd-common/Cargo.toml
@@ -16,7 +16,6 @@ crate-type = ["lib"]
 bench = false
 
 [dependencies]
-libdd-capabilities = { path = "../libdd-capabilities", version = "0.1.0" }
 anyhow = "1.0"
 futures = "0.3"
 futures-core = { version = "0.3.0", default-features = false }
