--- libdd-common/src/capabilities/mod.rs
@@ -1,6 +1,0 @@
-// Copyright 2026-Present Datadog, Inc. https://www.datadoghq.com/
-// SPDX-License-Identifier: Apache-2.0
-
-//! Re-exports of capability trait types for convenience.
-
-pub use libdd_capabilities::{HttpClientTrait, HttpError};
